:recycle: Refactor filter dropdown to be more similar to layout
