+++ src/components/transactionsV2/filters/filterContainer.test.js
@@ -4,10 +4,10 @@ import { shallow } from 'enzyme';
 import { spy } from 'sinon';
 import Input from 'react-toolbox/lib/input';
 import { PrimaryButtonV2 } from '../../toolbox/buttons/button';
-import FilterButton from './filterButton';
+import filterContainer from './filterContainer';
 import keyCodes from '../../../constants/keyCodes';
 
-describe('FilterButton', () => {
+describe('filterContainer', () => {
   it('should call saveFilters with expected Output', () => {
     const props = {
       saveFilters: spy(),
@@ -21,7 +21,7 @@ describe('FilterButton', () => {
       amountTo: '',
       message: 'test',
     };
-    const wrapper = shallow(<FilterButton {...props} />);
+    const wrapper = shallow(<filterContainer {...props} />);
     wrapper.find(Input).props().onChange('test');
     wrapper.find(PrimaryButtonV2).props().onClick();
     expect(props.saveFilters).to.be.calledWith(expectedValue);
@@ -33,7 +33,7 @@ describe('FilterButton', () => {
       customFilters: {},
     };
 
-    const wrapper = shallow(<FilterButton {...props} />);
+    const wrapper = shallow(<filterContainer {...props} />);
     expect(wrapper.state('showFilters')).to.be.been.equal(false);
     wrapper.find('.filterTransactions').simulate('click');
     expect(wrapper.state('showFilters')).to.be.been.equal(true);
@@ -54,7 +54,7 @@ describe('FilterButton', () => {
       amountTo: '',
       message: '',
     };
-    const wrapper = shallow(<FilterButton {...props} />);
+    const wrapper = shallow(<filterContainer {...props} />);
     wrapper.find(Input).props().onKeyDown({ keyCode: keyCodes.enter });
     expect(props.saveFilters).to.be.calledWith(expectedValue);
   });
